Smarter param handling in flo's FuncTarget
Increases the smarts of the `flo.targets.FuncTarget` class such
that the caller can mix and match params.  If a param appears to
be another target (it has a `future`), then the param will be
replaced by the result of its future, as before.  If a param does
not appear to be another target (no `future), then leave it alone.

This will allow for more natural usage where some params are
fixed and others are dependent on futures.