flo.channels.MapChannel
Adds a new flo Channel implementation that applies a mapper function
to each item from an input channel; the mapper function should return
an iterable, with each item emitted as the MapChannel's output
in order.  This allows a mapper function to effectively exclude
some things and explode other things out.