Flo CoGroupChannel more rigorous comparisons
The order of operands in comparison operations determines which
operand controls the comparison (via magical comparison methods).

Because we weren't ensuring that our initial MINIMUM/MAXIMUM
constants controlled the comparison, the flo.channels.CoGroupChannel
would fail if channels were keyed by things like datetime.

This fixes the problem to ensure they control the comparison,
and in so doing uses a new test that exercises CoGroupChannel with
datetime keys.