Refactor channelproperty tests for flexibility
I'll be adding a channelmethod decorator to address #41, which should
give a method-oriented interface equivalent to the property-oriented
interface we already have with channelproperty.

The testing for that will be nearly identical to the testing for the
channelproperty, except for how the channels are defined and accessed.

This refactors the existing tests (in flowz.test.manager_test) so the
channel definition and accessor logic are abstracted away from the
tests, so we can use specialization in subclasses to target the property
case versus the method case.